% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossover.R
\name{crossover}
\alias{crossover}
\title{Crossover Operator}
\usage{
crossover(gelite, gnelite, rc, p, pe, pm, n)
}
\arguments{
\item{gelite}{matrix with  elite chromosomes}

\item{gnelite}{matrix with nonelite chromosomes}

\item{rc}{crossover probability}

\item{p}{number of elements (chromosomes) in the population}

\item{pe}{number of elite  chromosomes}

\item{pm}{number of mutant chromossomes}

\item{n}{number of genes in the chromosomes associated with a solution}
}
\value{
\item{gnew}{= matrix pxn with chromosomes produced by crossover}
}
\description{
This function performs the uniform crossover between two chromosomes
}
\details{
crossover
}
\examples{
X<-popgen(10,10)
Y<-crossover(gelite=X[1:3,],gnelite=X[4:10,],rc=0.7,p=10,pe=3,pm=3,n=10)
}
\references{
1. Gonçalves, J.R. and Resende, M.G.C. (2011). Biased random-key genetic algorithms for combinatorial optimization,

2. Spears, W.M., DeJong, K.A.(1991). On the virtues of parameterized uniform crossover.
In: Proceedings of the Fourth International Conference on Genetic Algorithms, pp. 230b.
}
