% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_capitals.R
\docType{data}
\name{Data_capitals}
\alias{Data_capitals}
\title{File capitals}
\format{Data frame with 48 rows: (File capitals)}
\usage{
Data_capitals
}
\description{
Data Frame with 48 capitals of the US - TSP
}
\details{
Data_capitals
}
\references{
1. http://elib.zib.de/pub/mp-testdata/tsp/tsplib/tsp/index.html
}
\keyword{datasets}
