% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_MDP1.R
\docType{data}
\name{Data_MDP1}
\alias{Data_MDP1}
\title{File SOM-B_1_n100_m10}
\format{A distance Matrix 100 x 100: (File SOM-B_1_n100_m10)}
\usage{
Data_MDP1
}
\description{
Dataset with 100 points associated Maximum Diversity Problem
}
\details{
Data_MDP1
}
\references{
1. Silva, G.C., Ochi, L.S. e Martins, S.L. (2004). Experimental Comparison of Greedy Randomized Adaptive Search Procedures for the Maximum Diversity Problem. In Experimental and Efficient Algorithms, volume 3059 of Lecture Notes in Computer Science, p. 498b
}
\keyword{datasets}
