% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Decoder_ALLOC_SAMPLE.R
\name{Decoder_ALLOC_SAMPLE}
\alias{Decoder_ALLOC_SAMPLE}
\title{Decoder implemented to optimal allocation in stratified sampling}
\usage{
Decoder_ALLOC_SAMPLE(u, Nh)
}
\arguments{
\item{u}{chromossome with H position (number of strata)}

\item{Nh}{total Units in hth stratum (is necessary to define Exa1 = Nh)}
}
\value{
\item{nh}{= number of units in the sample in the hth stratum}
}
\description{
Builds a solution associated with sample sizes allocated by stratum
}
\details{
Decoder_ALLOC_SAMPLE
}
\examples{
H<-3
u<-popgen(n=H,p=1)
Nh<-c(212,84,61)
nh<-Decoder_ALLOC_SAMPLE(u,Nh)
}
