% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Decoder_HIMMELBLAUS.R
\name{Decoder_HIMMELBLAUS}
\alias{Decoder_HIMMELBLAUS}
\title{Decoder implemented to Himmelblau's function}
\usage{
Decoder_HIMMELBLAUS(u)
}
\arguments{
\item{u}{chromossome  with size = 2 (number of variables)}
}
\value{
\item{XY}{= vector with x e y values}
}
\description{
Calculates x e y values to HimmelBlau's Function
}
\details{
Decoder_HIMMELBLAUS
}
\examples{
u<-popgen(n=2,p=1)
XY<-Decoder_HIMMELBLAUS(u)
}
