% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Decoder_KNAPSACK.R
\name{Decoder_KNAPSACK}
\alias{Decoder_KNAPSACK}
\title{Decoder implemented to knapsack problem}
\usage{
Decoder_KNAPSACK(u)
}
\arguments{
\item{u}{chromossome  (with n positions associated with items)}
}
\value{
\item{xis}{= vector with 0 and 1s associated items of the knapsack}
}
\description{
Determines items that will be loaded in the knapsack.
}
\details{
Decoder_KNAPSACK
}
\examples{
nitems=7
u<-popgen(n=nitems,p=1)
xis<-Decoder_KNAPSACK(u)
}
\references{
1. Papadimitriou, C. H., and K. Steiglitz (1998). Combinatorial Optimization: Algorithms and Complexity. Dover Publications.
}
