% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brkga.R
\name{Decoder_MDP}
\alias{Decoder_MDP}
\title{Decoder implemented to Maximum Diversity Problem}
\usage{
Decoder_MDP(u, OD)
}
\arguments{
\item{u}{Chromossome}

\item{OD}{List with Ratio distance matrix and cardinality of M (is necessary to define Exa1 = OD)}
}
\value{
\item{Xm}{Elements of M}
}
\description{
Builds a feasible solution associated with a subset M of N
}
\details{
Decoder_MDP
}
