% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brkga.R
\name{Decoder_MSDCP}
\alias{Decoder_MSDCP}
\title{Decoder implemented to Minimum Sum Distance Clustering Problem}
\usage{
Decoder_MSDCP(u, k)
}
\arguments{
\item{u}{chromossome}

\item{k}{number of clusters (is necessary to define Exa1 = k)}
}
\value{
\item{cluster}{clustering vector}
}
\description{
Builds a clustering vector
}
\details{
Decoder_MSDCP
}
