% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brkga.R
\name{Fobj_ALLOC_SAMPLE}
\alias{Fobj_ALLOC_SAMPLE}
\title{Objective Function of the Optimal Allocation in Stratified Sampling}
\usage{
Fobj_ALLOC_SAMPLE(Sh2, nh, S)
}
\arguments{
\item{Sh2}{Variance by stratum}

\item{nh}{Number of units in the sample in the hth stratum}

\item{S}{List with Nh, Y and cvt (cv target)}
}
\value{
\item{n}{Sum of nh}
}
\description{
Calculates sum of sample sizes associated  with all strata
}
\details{
Fobj_ALLOC_SAMPLE
}
\references{
1. Brito, J. A. M.; Silva, P. L. N. ; Semaan, G. S. ; Maculan, N.(2015). Integer Programming Formulations Applied to Optimal Allocation in Stratified Sampling.
Survey Methodology, v. 41, p. 427-442.
}
