% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brkga.R
\name{Fobj_HIMMELBLAUS}
\alias{Fobj_HIMMELBLAUS}
\title{Himmelblau's function}
\usage{
Fobj_HIMMELBLAUS(Data = NULL, x)
}
\arguments{
\item{Data}{Function terms}

\item{x}{Point (x,y)}
}
\value{
\item{F}{Function Value}
}
\description{
Calculates value of the Himmelblau's function
}
\details{
Fobj_HIMMELBLAUS
}
\references{
1. Himmelblau, D. (1972). Applied Nonlinear Programming. McGraw-Hill.
}
