% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fobj_KNAPSACK.R
\name{Fobj_KNAPSACK}
\alias{Fobj_KNAPSACK}
\title{Objective function of the knapsack problem}
\usage{
Fobj_KNAPSACK(D, x, C)
}
\arguments{
\item{D}{matrix nx2 with weights (wi) and profits (li)}

\item{x}{items Knapsack}

\item{C}{capacity of knapsack (is necessary to define Exa2=C)}
}
\value{
\item{FP}{= profit associated items of knapscak}
}
\description{
Computes sum of profits based on itens allocated in knapsack
}
\details{
Fobj_KNAPSACK
}
\examples{
wi<-c(40,50,30,10,10,40,30)
li<-c(40,80,10,10,4,20,60)
x<-c(1,0,0,1,1,0,1)
C<-100
Datalw<-cbind(li,wi)
FP<-Fobj_KNAPSACK(Datalw,x,C)
}
\references{
1. Papadimitriou, C. H., and K. Steiglitz (1998). Combinatorial Optimization: Algorithms and Complexity. Dover Publications.

2. Martello, S. and Toth, P. (1990). Knapsack Problems - Algorithms and Computer Implementations. Wiley.
}
