% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brkga.R
\name{Fobj_MDP}
\alias{Fobj_MDP}
\title{Objective Function of the Maximum Diversity Problem}
\usage{
Fobj_MDP(D, Xm)
}
\arguments{
\item{D}{distance matrix}

\item{Xm}{Elements of M}
}
\value{
\item{SD}{Sum all distances associated with M}
}
\description{
Calculates the sum of all distances associated with M
}
\details{
Fobj_MDP
}
\references{
1. Martí, R., Gallego, M., Duarte, A. e Pardo, E. G. (2013) Heuristics and metaheuristics for the maximum diversity problem. Journal of Heuristics 19(4): 591-615.

2. Silva, G.C., Ochi, L.S. e Martins, S.L. (2004).
Experimental Comparison of Greedy Randomized Adaptive Search Procedures for the Maximum Diversity Problem.
In Experimental and Efficient Algorithms, volume 3059 of Lecture Notes in Computer Science, p. 498b-512.
Springer Berlin / Heidelberg.
}
