% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fobj_MEDOID.R
\name{Fobj_MEDOID}
\alias{Fobj_MEDOID}
\title{Objective function of the k-medoids problem}
\usage{
Fobj_MEDOID(D, x, k)
}
\arguments{
\item{D}{distance matrix}

\item{x}{medoids}

\item{k}{number of clusters (is necessary to define Exa2=k)}
}
\value{
\item{S}{= sum of distances}
}
\description{
Computes the sum of the distances for each object from near
object associated with medoid
}
\details{
Fobj_MEDOID
}
\examples{
D<-as.matrix(dist(iris[,1:4]))
N<-nrow(iris)
k<-3 #clusters
x<-sample(N,k)  #selected medoids
fobj<-Fobj_MEDOID(D,x,k)
}
\references{
1. Kaufman L. e Rousseeuw P.J. (1989). Finding Groups in Data. An Introduction to Cluster Analysis.   Wiley-Interscience Publication.

2. Zhang, Q. and Couloigner, I. (2005). A New Efficient k-medoid Algorithm for Spatial Clustering. Lecture Notes in Computer Science, v3482, 181-189.
}
