% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fobj_MSDCP.R
\name{Fobj_MSDCP}
\alias{Fobj_MSDCP}
\title{Objective function of the minimum sum distance clustering problem}
\usage{
Fobj_MSDCP(D, cluster, k)
}
\arguments{
\item{D}{distance Matrix}

\item{cluster}{clustering Vector}

\item{k}{number of clusters (is necessary to define Exa2=k)}
}
\value{
\item{SD}{= sum of distances}
}
\description{
Fobj_MSDCP
}
\examples{
k<-3 #Clusters
Distance<-as.matrix(dist(iris[,1:4]))
N<-nrow(iris)
u<-popgen(n=N,p=1)
cluster<-Decoder_MSDCP(u,k)
SD<-Fobj_MSDCP(Distance,cluster,k)
}
\references{
1. Friedman, J. H. and Meulman J.J.  (2004). Clustering objects on subsets of attributes.
#' Journal Royal Statistics Society B, Part 4,66:815-849.

2. Hansen, P. and Jaumard, B. (1997). Cluster Analysis and Mathematical Programming.
#'Mathematical  Programming, 79:191-215.

3. Rao, M.R. (1971). Cluster Analysis and Mathematical Programming.
Journal of American Statistical Association, 66:622-626.
}
