% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fobj_TSP.R
\name{Fobj_TSP}
\alias{Fobj_TSP}
\title{Objective function of the travelling salesman problem}
\usage{
Fobj_TSP(D, x)
}
\arguments{
\item{D}{Distance matrix}

\item{x}{Route}
}
\value{
\item{sD}{= sum of distances}
}
\description{
Computes the sum of the distances based on the route x and
D matrix (matrix with the distances between cities)
}
\details{
Fobj_TSP
}
\examples{
data(Data_capitals)
D<-as.matrix(dist(Data_capitals))
Ncities<-nrow(Data_capitals)
x<-sample(Ncities,Ncities)
fobj<-Fobj_TSP(D,x)
}
\references{
1. Applegate David L. et al. (2006). The travelling salesman problem: a computational Study. Princeton: Princeton University Press.

2. Garey, M. and Johnson, D. S (1990). Computer and Intractibility: A guide to the Theory of NPCompleteness, Freeman, San Francisco.
}
