% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getChromosomeFitness}
\alias{getChromosomeFitness}
\title{Get fitness from chromosome}
\usage{
getChromosomeFitness(chromosome, Distances, m)
}
\arguments{
\item{\code{chromosome}}{Chromosome}

\item{\code{Distances}}{Distance matrix}

\item{\code{m}}{Size of tour}
}
\value{
A double value representing the chromosome fitness
}
\description{
Get fitness from chromosome
}
\details{
Get fitness using chromossome, m and distance matrix
}
