% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tsp_brkga}
\alias{tsp_brkga}
\title{Execute a MDP solution search using BRKGA}
\usage{
tsp_brkga(instanceFile)
}
\arguments{
\item{MAX_TIME}{Max time of execution in seconds}

\item{p}{Size of population}

\item{pe}{Fraction of population to be the elite-set}

\item{pm}{Fraction of population to be replaced by mutants}

\item{rhoe}{Frobability that offspring inherit an allele from elite parent}

\item{K}{number of independent populations}

\item{MAXT}{number of threads for parallel decoding}

\item{X_INTVL}{Exchange best individuals at every 100 generations}

\item{X_NUMBER}{Exchange top 2 best}

\item{MAX_GENS}{Max number of generations}

\item{SEM_MELHORA}{Pergunte para Geiza}

\item{verbose}{Level of output information}

\item{rngSeed}{Seed to the random number generator}
}
\value{
A numeric vector of random values
}
\description{
Execute a MDP solution search using BRKGA
}
\details{
Escrever um detalhamento bem legal e super bacana
 aqui.
}
\seealso{
\code{\link{api-usage}} and \url{https://github.com/milkway/brkga}
}
