% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brkga.R
\name{Decoder_ALLOC_SAMPLE}
\alias{Decoder_ALLOC_SAMPLE}
\title{Decoder implemented to Optimal Allocation in Stratified Sampling}
\usage{
Decoder_ALLOC_SAMPLE(u, Nh)
}
\arguments{
\item{u}{chromossome}

\item{Nh}{Total Units in hth stratum (is necessary to define Exa1 = Nh)}
}
\value{
\item{nh}{Number of units in the sample in the hth stratum}
}
\description{
Builds a solution associated with sample sizes allocated by stratum
}
\details{
Decoder_ALLOC_SAMPLE
}
