% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brkga.R
\name{Decoder_HIMMELBLAUS}
\alias{Decoder_HIMMELBLAUS}
\title{Decoder implemented to Himmelblau's function}
\usage{
Decoder_HIMMELBLAUS(u)
}
\arguments{
\item{u}{chromossome}
}
\value{
\item{XY}{Vector with x e y values}
}
\description{
Calculates x e y values to HimmelBlau's Function
}
\details{
Decoder_HIMMELBLAUS
}
