% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brkga.R
\name{Decoder_KNAPSACK}
\alias{Decoder_KNAPSACK}
\title{Decoder implemented to Knapsack Problem}
\usage{
Decoder_KNAPSACK(u)
}
\arguments{
\item{u}{chromossome}
}
\value{
\item{xis}{Vector with 0 and 1s associated items of the knapsack}
}
\description{
Determines items that will be loaded in the knapsack.
}
\details{
Decoder_Knapsack
}
\references{
1. Papadimitriou, C. H., and K. Steiglitz (1998). Combinatorial Optimization: Algorithms and Complexity. Dover Publications.
}
