% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Decoder_MDP.R
\name{Decoder_MDP}
\alias{Decoder_MDP}
\title{Decoder implemented to maximum diversity problem}
\usage{
Decoder_MDP(u, OD)
}
\arguments{
\item{u}{Chromossome}

\item{OD}{List with Ratio distance matrix and cardinality of M (is necessary to define Exa1 = OD)}
}
\value{
\item{Xm}{= elements of M}
}
\description{
Builds a feasible solution associated with a subset M of N
}
\details{
Decoder_MDP
}
\examples{
data(Data_MDP1)
D<-Data_MDP1
N<-dim(D)[1]
M<-10
RDi<-apply(as.matrix(1:N),1,function(i) sum(D[i,]))/sum((D[upper.tri(D)]))
u<-popgen(n=N,1)
OD<-list(M=M,RDi=RDi)
Xm<-Decoder_MDP(u,OD)
}
