% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brkga.R
\name{Decoder_MEDOID}
\alias{Decoder_MEDOID}
\title{Decoder implemented to k-medoids Problem}
\usage{
Decoder_MEDOID(u, k)
}
\arguments{
\item{u}{chromossome}

\item{k}{number of clusters (is necessary to define Exa1 = k)}
}
\value{
\item{medoids}{Vector with k-medoids}
}
\description{
Determines k objects associated with medoids
}
\details{
Decoder_MEDOID

k-medoids Problem is fully described in Kaufman and Rousseeuw (1989)
}
\references{
1. Kaufman L. e Rousseeuw P.J. (1989). Finding Groups in Data. An Introduction to Cluster Analysis.   Wiley-Interscience Publication.

2. Zhang, Q. and Couloigner, I. (2005). A New Efficient k-medoid Algorithm for Spatial Clustering. Lecture Notes in Computer Science, v3482, 181-189.
}
