% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Decoder_MSDCP.R
\name{Decoder_MSDCP}
\alias{Decoder_MSDCP}
\title{Decoder implemented to minimum sum distance clustering problem}
\usage{
Decoder_MSDCP(u, k)
}
\arguments{
\item{u}{chromossome}

\item{k}{number of clusters (is necessary to define Exa1 = k)}
}
\value{
\item{cluster}{ = clustering vector}
}
\description{
Builds a clustering vector
}
\details{
Decoder_MSDCP
}
\examples{
N<-nrow(iris)
k<-3
u<-popgen(n=N,p=1)
cluster<-Decoder_MSDCP(u,k)
}
