% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Decoder_TSP.R
\name{Decoder_TSP}
\alias{Decoder_TSP}
\title{Decoder implemented to tsp problem}
\usage{
Decoder_TSP(u)
}
\arguments{
\item{u}{chromossome (n positions)}
}
\value{
\item{x}{= Route with sequence of cities}
}
\description{
Builds a feasible solution associated a route (order of travel of n cities).
}
\details{
Decoder_TSP

TSP
}
\examples{
data(Data_capitals)
Ncities<-nrow(Data_capitals)
u<-popgen(n=Ncities,p=1)
x<-Decoder_TSP(u)
}
\references{
1. Applegate David L. et al. (2006). The Travelling Salesman Problem: a computational study. Princeton: Princeton University Press.

2. Garey, M. and Johnson, D. S (1990). Computer and Intractibility: A guide to the Theory of NPCompleteness, Freeman, San Francisco.

3. Kumar, K.(2012). Traveling Salesman Problem (TSP): A Comparative Analysis: Optimization Techniques for Traveling Salesman Problem. LAP LAMBERT Academic Publishing
}
