% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fobj_HIMMELBLAUS.R
\name{Fobj_HIMMELBLAUS}
\alias{Fobj_HIMMELBLAUS}
\title{Himmelblau's function}
\usage{
Fobj_HIMMELBLAUS(Data = NULL, x)
}
\arguments{
\item{Data}{Function terms}

\item{x}{Point (x,y)}
}
\value{
\item{F}{= function Value}
}
\description{
Calculates value of the Himmelblau's function
}
\details{
Fobj_HIMMELBLAUS
}
\examples{
F<-Fobj_HIMMELBLAUS(x=c(1.1,1.5))
}
\references{
1. Himmelblau, D. (1972). Applied Nonlinear Programming. McGraw-Hill.
}
