% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brkga.R
\name{Fobj_KNAPSACK}
\alias{Fobj_KNAPSACK}
\title{Objective Function of the Knapsack Problem}
\usage{
Fobj_KNAPSACK(D, x, C)
}
\arguments{
\item{D}{Distance matrix}

\item{x}{Items Knapsack}

\item{C}{Capacity of knapsack}
}
\value{
\item{FP}{Profit associated items of knapscak}
}
\description{
Computes Sum of Profits based on itens allocated in knapsack
}
\details{
Fobj_KNAPSACK
}
\references{
1. Papadimitriou, C. H., and K. Steiglitz (1998). Combinatorial Optimization: Algorithms and Complexity. Dover Publications.

2. Martello, S. and Toth, P. (1990). Knapsack Problems - Algorithms and Computer Implementations. Wiley.
}
