% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fobj_MDP.R
\name{Fobj_MDP}
\alias{Fobj_MDP}
\title{Objective function of the maximum diversity problem}
\usage{
Fobj_MDP(D, Xm)
}
\arguments{
\item{D}{distance matrix}

\item{Xm}{elements of M}
}
\value{
\item{SD}{= sum all distances associated with M}
}
\description{
Calculates the sum of all distances associated with M
}
\details{
Fobj_MDP
}
\examples{
data(DataMDP1)
D<-Data_MDP1
N<-dim(D)[1]
M<-10
xm<-sample(N,M)
SD<-Fobj_MDP(D,xm)
}
\references{
1. Martí, R., Gallego, M., Duarte, A. e Pardo, E. G. (2013) Heuristics and metaheuristics for the maximum diversity problem. Journal of Heuristics 19(4): 591-615.

2. Silva, G.C., Ochi, L.S. e Martins, S.L. (2004).
Experimental Comparison of Greedy Randomized Adaptive Search Procedures for the Maximum Diversity Problem.
In Experimental and Efficient Algorithms, volume 3059 of Lecture Notes in Computer Science, p. 498b-512.
Springer Berlin / Heidelberg.
}
