% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brkga.R
\name{Fobj_MEDOID}
\alias{Fobj_MEDOID}
\title{Objective Function of the k-medoids Problem}
\usage{
Fobj_MEDOID(D, x, k)
}
\arguments{
\item{D}{Distance matrix}

\item{x}{Medoids}

\item{k}{Number of Clusters}
}
\value{
\item{Soma}{Sum of distances total}
}
\description{
Computes the sum of the distances for each object from near
object associated with medoid
}
\details{
Fobj_MEDOID
}
\references{
1. Kaufman L. e Rousseeuw P.J. (1989). Finding Groups in Data. An Introduction to Cluster Analysis.   Wiley-Interscience Publication.

2. Zhang, Q. and Couloigner, I. (2005). A New Efficient k-medoid Algorithm for Spatial Clustering. Lecture Notes in Computer Science, v3482, 181-189.
}
