% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brkga.R
\name{Fobj_MSDCP}
\alias{Fobj_MSDCP}
\title{Objective Function of the Minimum Sum Distance Clustering Problem}
\usage{
Fobj_MSDCP(D, cluster, k)
}
\arguments{
\item{D}{Distance Matrix}

\item{cluster}{Clustering Vector}

\item{k}{Number of clusters}
}
\value{
\item{SD}{Sum Distances}
}
\description{
Fobj_MSDCP
}
\references{
1. Friedman, J. H. and Meulman J.J.  (2004). Clustering objects on subsets of attributes.
#' Journal Royal Statistics Society B, Part 4,66:815-849.

2. Hansen, P. and Jaumard, B. (1997). Cluster Analysis and Mathematical Programming.
#'Mathematical  Programming, 79:191-215.

3. Rao, M.R. (1971). Cluster Analysis and Mathematical Programming.
Journal of American Statistical Association, 66:622-626.
}
