% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brkga.R
\name{Fobj_TSP}
\alias{Fobj_TSP}
\title{Objective Function of the Travelling Salesman Problem}
\usage{
Fobj_TSP(D, x)
}
\arguments{
\item{D}{Distance matrix}

\item{x}{Route}
}
\value{
\item{sD}{Sum of distances}
}
\description{
Computes the sum of the distances based on the route x and
D matrix (matrix with the distances between cities)
}
\details{
Fobj_TSP
}
\references{
1. Applegate David L. et al. (2006). The travelling salesman problem: a computational Study. Princeton: Princeton University Press.

2. Garey, M. and Johnson, D. S (1990). Computer and Intractibility: A guide to the Theory of NPCompleteness, Freeman, San Francisco.
}
