% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mdp_brkga}
\alias{mdp_brkga}
\title{Execute a MDP solution search using BRKGA}
\usage{
mdp_brkga(instanceFile, MAX_TIME = 10L, p = 500L, pe = 0.2,
  pm = 0.1, rhoe = 0.7, K = 1L, MAXT = 8L, X_INTVL = 100L,
  X_NUMBER = 2L, MAX_GENS = 1000L, SEM_MELHORA = 0L,
  verbose = FALSE, rngSeed = 0L)
}
\arguments{
\item{instanceFile}{Path to file containing the instance}

\item{MAX_TIME}{Max time of execution in seconds}

\item{p}{Size of population}

\item{pe}{Fraction of population to be the elite-set}

\item{pm}{Fraction of population to be replaced by mutants}

\item{rhoe}{Frobability that offspring inherit an allele from elite parent}

\item{K}{number of independent populations}

\item{MAXT}{number of threads for parallel decoding}

\item{X_INTVL}{Exchange best individuals at every 100 generations}

\item{X_NUMBER}{Exchange top 2 best}

\item{MAX_GENS}{Max number of generations}

\item{SEM_MELHORA}{Pergunte para Geiza}

\item{verbose}{Level of output information}

\item{rngSeed}{Seed to the random number generator}
}
\value{
A numeric vector of random values
}
\description{
Execute a MDP solution search using BRKGA
}
\details{
Escrever um detalhamento bem legal e super bacana
 aqui.
}
\seealso{
\code{\link{api-usage}} and \url{https://github.com/milkway/brkga}
}
