% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brkga.R
\name{popgen}
\alias{popgen}
\title{Population Generation}
\usage{
popgen(n, p)
}
\arguments{
\item{n}{Number of genes in the chromosome associated with a solution}

\item{p}{Number of elements (chromosomes) in the population}
}
\value{
\item{pop}{The population is stored in the p x n real-valued matrix}
}
\description{
Generates p vectors of random-keys
}
\details{
popgen
}
\references{
1. Gonçalves, J.R. and Resende, M.G.C. (2011). Biased random-key genetic algorithms for combinatorial optimization,
}
